Remove variable shadowing from tensorpipe lambda
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39126
futureResponseMessage is shadowed in the pipeWrite lambda which
creates some confusion, since it is used in the initial error handling but then
a future of the same name is created when marking the future as completed. This
change removes this by getting rid of the futureResponseMessage capture,
instead capturing the message id. This change also makes it so that we don't
need to copy it into the lambda.

Differential Revision: [D22127398](https://our.internmc.facebook.com/intern/diff/D22127398/)
ghstack-source-id: 106197169